Song lyrics - iOS

+++ BMM.Api/Implementation/Models/Transcription.cs
@@ -3,7 +3,7 @@ namespace BMM.Api.Implementation.Models;
 public class Transcription
 {
     public int Id { get; set; }
-    public decimal Start { get; set; }
-    public decimal End { get; set; }
+    public double Start { get; set; }
+    public double End { get; set; }
     public string Text { get; set; }
 }
